apiextensions: cleanup test/integration helpers
